refactor(minifySvg): rewrite to typescript
