+++ src/index.ts
@@ -69,7 +69,7 @@ const modules = {
     minifyCss: () => import('./_modules/minifyCss.mjs'),
     minifyJs: () => import('./_modules/minifyJs.mjs'),
     minifyJson: () => interop(import('./_modules/minifyJson.js')),
-    minifySvg: () => import('./_modules/minifySvg.mjs'),
+    minifySvg: () => interop(import('./_modules/minifySvg.js')),
     minifyUrls: () => import('./_modules/minifyUrls.mjs'),
     normalizeAttributeValues: () => interop(import('./_modules/normalizeAttributeValues')),
     removeAttributeQuotes: () => interop(import('./_modules/removeAttributeQuotes.js')),
